pyproject.toml
[project]
name = "truehair-mvp-2"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.66.0",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "boto3>=1.38.0",
    "flask-migrate>=4.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-flask>=1.3.0",
    "ruff>=0.15.5",
    "pre-commit>=4.1.0",
]
